Failed attempt at improving shadows. DO NOT MERGE
After digging into shadows some more we cannot implement shadows by hiding the
views without significant work. A better method is probably to make everything
except the controller unaware of shadow targets on connections and then have the
controller move the shadow block between the target and the shadow target and
create/destroy its views.

Leaving this CL in as pieces of it may be useful in that future change.